[Refactor] Pull out Invite Modal and All Members table into own files (#5441)
## Summary
We are going to be adding logic to the `WorkspaceTeam` page to support
different views (i.e. "Pending invites"), but the file is getting large
and crowded. Pull out the invite modal and "All members" logic into
their own files to maintain and reuse if necessary.

## How did you test this change?
1. Add the necessary API keys to send emails or update the
[sendEmailInvite](https://github.com/highlight/highlight/pull/5441/files#diff-fd55b7db8c5cb2f6c62f0a50cbdb7b1c603ef753a6bc1122ab15eeee7cc8aed7R66)
to use `finally` instead of `then`, so it will display the success
message.
2. Open the modal on the"User management" page (`/w/1/team`) and send an
invite to a team member
- [x] Confirm no changes in behavior in modal (email "sent", toast
displayed, etc)
- [x] Confirm no visual changes in modal
3) Close modal
- [ ] Confirm no changes in behavior on "User management" page
- [ ] Confirm no visual changes on page



https://github.com/highlight/highlight/assets/17744174/12abf9c5-7005-4606-8f23-c01cc6837b5c

## Are there any deployment considerations?
N/A

---------
