+++ frontend/src/__generated/scss/pages/WorkspaceTeam/WorkspaceTeam.module.scss.js
@@ -1,1 +1,1 @@
-var t={subTitle:"_subTitle_eqxhx_1",membersTitleContainer:"_membersTitleContainer_eqxhx_5",membersTitle:"_membersTitle_eqxhx_5",memberCard:"_memberCard_eqxhx_13",memberCardWrapper:"_memberCardWrapper_eqxhx_20",userDetails:"_userDetails_eqxhx_26",name:"_name_eqxhx_33",email:"_email_eqxhx_40",role:"_role_eqxhx_48",header:"_header_eqxhx_54",buttonRow:"_buttonRow_eqxhx_59",emailInput:"_emailInput_eqxhx_66",inviteButton:"_inviteButton_eqxhx_70",boxSubTitle:"_boxSubTitle_eqxhx_74",removeTeamMemberButton:"_removeTeamMemberButton_eqxhx_79",titleContainer:"_titleContainer_eqxhx_85",hr:"_hr_eqxhx_98"};export{t as default};
+var e={titleContainer:"_titleContainer_1oin8_1",subTitle:"_subTitle_1oin8_1"};export{e as default};
