+++ frontend/src/__generated/scss/pages/WorkspaceTeam/components/AllMembers.module.scss.js
@@ -1,0 +1,1 @@
+var m={membersTitleContainer:"_membersTitleContainer_1a35s_1",membersTitle:"_membersTitle_1a35s_1",memberCard:"_memberCard_1a35s_9",memberCardWrapper:"_memberCardWrapper_1a35s_16",email:"_email_1a35s_22",role:"_role_1a35s_30",removeTeamMemberButton:"_removeTeamMemberButton_1a35s_36"};export{m as default};
