+++ frontend/src/pages/WorkspaceTeam/WorkspaceTeam.module.scss
@@ -1,87 +1,3 @@
-.subTitle {
-	font-size: 18px;
-}
-
-.membersTitleContainer {
-	margin-bottom: var(--size-xSmall);
-}
-
-.membersTitle {
-	font-weight: 500 !important;
-}
-
-.memberCard {
-	align-items: center;
-	column-gap: var(--size-small);
-	display: flex;
-	flex-direction: row;
-}
-
-.memberCardWrapper {
-	margin-right: var(--size-xxLarge) !important;
-	max-width: 700px;
-	width: 100%;
-}
-
-.userDetails {
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	overflow: hidden;
-}
-
-.name {
-	color: var(--text-primary);
-	overflow: hidden;
-	text-overflow: ellipsis;
-	white-space: nowrap;
-}
-
-.email {
-	color: var(--color-gray-500);
-	font-size: 13px;
-	overflow: hidden;
-	text-overflow: ellipsis;
-	white-space: nowrap;
-}
-
-.role {
-	color: var(--color-gray-500);
-	font-size: 13px;
-	justify-content: flex-end;
-}
-
-.header {
-	column-gap: var(--size-small);
-	display: flex;
-}
-
-.buttonRow {
-	align-items: center;
-	column-gap: var(--size-medium);
-	display: grid;
-	grid-template-columns: 1fr 120px;
-}
-
-.emailInput {
-	height: 40px;
-}
-
-.inviteButton {
-	width: 100%;
-}
-
-.boxSubTitle {
-	color: var(--color-gray-500) !important;
-	margin-top: 0 !important;
-}
-
-.removeTeamMemberButton {
-	flex-shrink: 0;
-	height: fit-content;
-	margin-left: auto;
-}
-
 .titleContainer {
 	align-items: flex-start;
 	display: flex;
@@ -94,10 +10,4 @@
 		margin-bottom: 0 !important;
 		margin-top: var(--size-xSmall) !important;
 	}
-}
-
-.hr {
-	border: 0;
-	border-top: 1px solid var(--color-gray-300);
-	margin: var(--size-large) 0;
 }
