+++ frontend/src/pages/WorkspaceTeam/WorkspaceTeam.module.scss.d.ts
@@ -1,17 +1,2 @@
-export const boxSubTitle: string
-export const buttonRow: string
-export const email: string
-export const emailInput: string
-export const header: string
-export const hr: string
-export const inviteButton: string
-export const memberCard: string
-export const memberCardWrapper: string
-export const membersTitle: string
-export const membersTitleContainer: string
-export const name: string
-export const removeTeamMemberButton: string
-export const role: string
 export const subTitle: string
 export const titleContainer: string
-export const userDetails: string
