+++ frontend/src/pages/WorkspaceTeam/components/AllMembers.module.scss
@@ -1,0 +1,40 @@
+.membersTitleContainer {
+	margin-bottom: var(--size-xSmall);
+}
+
+.membersTitle {
+	font-weight: 500 !important;
+}
+
+.memberCard {
+	align-items: center;
+	column-gap: var(--size-small);
+	display: flex;
+	flex-direction: row;
+}
+
+.memberCardWrapper {
+	margin-right: var(--size-xxLarge) !important;
+	max-width: 700px;
+	width: 100%;
+}
+
+.email {
+	color: var(--color-gray-500);
+	font-size: 13px;
+	overflow: hidden;
+	text-overflow: ellipsis;
+	white-space: nowrap;
+}
+
+.role {
+	color: var(--color-gray-500);
+	font-size: 13px;
+	justify-content: flex-end;
+}
+
+.removeTeamMemberButton {
+	flex-shrink: 0;
+	height: fit-content;
+	margin-left: auto;
+}
