+++ frontend/src/pages/WorkspaceTeam/components/AllMembers.module.scss.d.ts
@@ -1,0 +1,7 @@
+export const email: string
+export const memberCard: string
+export const memberCardWrapper: string
+export const membersTitle: string
+export const membersTitleContainer: string
+export const removeTeamMemberButton: string
+export const role: string
